+++ plugins/http2_adapter/test/http2_test.dart
@@ -49,4 +49,31 @@ void main() {
     final res = await dio.post('post', data: 'TEST');
     expect(res.data.toString(), contains('TEST'));
   });
+
+  test('request without network and restore', () async {
+    bool needProxy = true;
+    final dio = Dio()
+      ..options.baseUrl = 'https://httpbun.com/'
+      ..httpClientAdapter = Http2Adapter(ConnectionManager(
+        idleTimeout: Duration(milliseconds: 10),
+        onClientCreate: (uri, settings) {
+          if (needProxy) {
+            // first request use bad proxy to simulate network error
+            settings.proxy = Uri.parse('http://localhost:1234');
+            needProxy = false;
+          } else {
+            // remove proxy to restore network
+            settings.proxy = null;
+          }
+        },
+      ));
+    try {
+      // will throw SocketException
+      await dio.post('post', data: 'TEST');
+    } on DioException {
+      // ignore
+    }
+    final res = await dio.post('post', data: 'TEST');
+    expect(res.data.toString(), contains('TEST'));
+  });
 }