+++ plugins/http2_adapter/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-*None.*
+- Fix cached `initFuture` not remove when throw exception.
 
 ## 2.3.0